List: add status indicator
Adds extension to render appearance for status indicator fields. It
will render if an 'indicator' option is included in the field
attributes. Each relevant state is then mapped either to 'on' ->
'Green' appearance, or 'off' -> 'Red' appearance. If the item's state
is not mapped, then it will default to a 'Gray' appearance.

Sample:

fields: {
  [...],

  state: {
    label: 'Status',
    indicator: {
      'Running': 'on',
      'Stopped': 'off',
      'Error': 'off'
    }
  }
}